requirements.txt
boto3==1.34.144
streamlit==1.38.0
python-dotenv==1.0.1
Pillow==10.4.0